refactor: group the generic and Firefly importers under transaction imports
