--- backend/app/services/data_imports/__init__.py
@@ -1,6 +1,0 @@
-"""Importers for data exported from external finance apps
-
-Each supported source lives in its own subpackage and converts that app's
-export format into Lumina transactions using the shared transaction import
-helpers
-"""
