+++ backend/app/services/transactions/imports/generic/__init__.py
@@ -1,0 +1,5 @@
+"""Generic CSV transaction importer
+
+Converts rows a user mapped column by column into transactions, one row
+becoming one transaction on the account the row was mapped to
+"""
