+++ backend/app/services/transactions/imports/generic/lookup_helpers.py
@@ -11,12 +11,12 @@ from app.models.merchant import Merchant
 from app.models.tag import Tag
 from app.models.user import User
 from app.schemas.transaction import TransactionImportRequest
-from app.services.transactions.imports.accounts import get_or_create_import_accounts_by_source
-from app.services.transactions.imports.categories import get_or_create_import_categories_by_source
-from app.services.transactions.imports.currencies import get_import_currencies_by_code
-from app.services.transactions.imports.merchants import get_personal_import_merchants_by_name
-from app.services.transactions.imports.stats import ImportStats
-from app.services.transactions.imports.tags import get_personal_import_tags_by_name
+from app.services.transactions.imports.shared.accounts import get_or_create_import_accounts_by_source
+from app.services.transactions.imports.shared.categories import get_or_create_import_categories_by_source
+from app.services.transactions.imports.shared.currencies import get_import_currencies_by_code
+from app.services.transactions.imports.shared.merchants import get_personal_import_merchants_by_name
+from app.services.transactions.imports.shared.stats import ImportStats
+from app.services.transactions.imports.shared.tags import get_personal_import_tags_by_name
 
 
 @dataclass
