+++ backend/app/services/transactions/imports/shared/__init__.py
@@ -1,0 +1,7 @@
+"""Machinery both import sources build on
+
+Resolves the accounts, categories, currencies, merchants, and tags an
+import references, either matching what the user already has or creating
+what the payload describes, and carries the running statistics an import
+reports back
+"""
